correct type

+++ display_image_it8951.py
@@ -116,7 +116,7 @@ def _fetch_image_from_urlfile(url, resizeWidth, resizeHeight, fill, scale):
             scaled_image = ImageOps.scale(image, scale_factor, resample=Image.Resampling.LANCZOS)
             return scaled_image.crop((0, 0, resizeWidth, resizeHeight))
         else:
-            padded_image = ImageOps.pad(image, (resizeWidth * scale, resizeHeight * scale), color=(255,255,255), method=Image.Resampling.LANCZOS, centering=(0.5, 0.5))
+            padded_image = ImageOps.pad(image, (int(resizeWidth * scale), int(resizeHeight * scale)), color=(255,255,255), method=Image.Resampling.LANCZOS, centering=(0.5, 0.5))
             return padded_image.crop((0, 0, scaled_image.width, resizeHeight))
 
 def do_webpage_display(url, width, height, virtual, rotate, mirror, fill):
